Add macvtap-cni production code
The ConfigureMacvtap interface code is located in the
`netlink` util pkg.

The following helper functions are also made available on
the `netlink` pkg.
  - configureArp
  - renameMacvtapIface

All the configuration is performed within a single netns.Do
instruction.

Interface renaming is performed in the target netns, thus
reducing the risk of interface name collision.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>