Add cni plugin (#4)
* Add macvtap-cni production code

The ConfigureMacvtap interface code is located in the
`netlink` util pkg.

The following helper functions are also made available on
the `netlink` pkg.
  - configureArp
  - renameMacvtapIface

All the configuration is performed within a single netns.Do
instruction.

Interface renaming is performed in the target netns, thus
reducing the risk of interface name collision.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>

* Use init containers to install the cni plugin

The init-container installs the macvtap-cni plugin into the
host, and afterwards, the macvtap container listens to device
plugin requests, and operates on them.

This way, we can achieve macvtap cni & dp co-existence within the
same container.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>

* Add unit test to check the correct behavior of the CNI plugin

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>

* workaround: remove flag breaking unit tests

The -logtostderr flag is somehow preventing the cni unit tests
from running.

Can't figure this one out; this is just to show I'm not insane.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>